requirements.txt
Pillow==5.1.0
numpy==1.14.2
